move portfolios.py to models package
