--- backend/ops_api/ops/portfolio/urls.py
@@ -1,11 +1,0 @@
-from flask import Blueprint
-
-bp_portfolio_status = Blueprint(
-    "portfolio_status", __name__, url_prefix="/ops/portfolio_status"
-)
-# bp_portfolio_status.add_url_rule("/", view_func=views.portfolio_status_list)
-# bp_portfolio_status.add_url_rule("/<int:pk>/", view_func=views.get_portfolio_status)
-
-bp_division = Blueprint("division", __name__, url_prefix="/ops/division")
-# bp_division.add_url_rule("/", view_func=views.division_list)
-# bp_division.add_url_rule("/<int:pk>/", view_func=views.get_division)
