+++ backend/ops_api/ops/summary/urls.py
@@ -4,7 +4,7 @@ from flask import request
 from flask import Response
 from ops.can.utils import get_can_funding_summary as get_can_funding
 from ops.models.cans import CAN
-from ops.portfolio.models import Portfolio
+from ops.models.portfolios import Portfolio
 from ops.portfolio.utils import get_total_funding as get_portfolio_funding
