+++ backend/ops_api/ops/urls.py
@@ -1,9 +1,9 @@
 from ops.models.cans import BudgetLineItem
 from ops.models.cans import CAN
 from ops.models.cans import CANFiscalYear
-from ops.portfolio.models import Division
-from ops.portfolio.models import Portfolio
-from ops.portfolio.models import PortfolioStatus
+from ops.models.portfolios import Division
+from ops.models.portfolios import Portfolio
+from ops.models.portfolios import PortfolioStatus
 from ops.resources.auth import AuthLoginAPI
 from ops.resources.auth import AuthRefreshAPI
 from ops.resources.budget_line_items import BudgetLineItemsItemAPI
