+++ backend/ops_api/tests/ops/portfolio/test_portfolio_status.py
@@ -1,4 +1,4 @@
-from ops.portfolio.models import PortfolioStatus
+from ops.models.portfolios import PortfolioStatus
 import pytest
